✨🎨🏗🛠 improve code structure ✨🎨🏗🛠
download_urls.py:

        search_in_youtube:
Improved logging information in the console related to
searching for tracks on YouTube.
Attempting to retrieve a cached result first, and if not
available, caching a new result.

        playlist_downloader:
Modified to download each album of an artist to a separate
folder within the artist's folder.

    spots.py:
Changed the first step to navigate to the Music folder to load
the cache into memory.
        Now using the port number set in the environment.

    server/download_music.py:
The /download/artist|playlist endpoint now accepts a query
parameter to indicate an artist search query.